requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
xlsxwriter==3.2.9
openpyxl>=3.0.0
boto3
pypdf
celery
redis
